(test) Use redux-mock-store and moxios to test actions
